Rename BitcoinAccount to BlockinfoResponse
