+++ app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/client/BlockonomicsRequest.java
@@ -1,0 +1,4 @@
+package io.github.jangerhard.BitcoinWalletTracker.client;
+
+public class BlockonomicsRequest {
+}
